mexendo

+++ investimento/dividendos.hs
@@ -1,5 +1,5 @@
 module Investimento.Dividendos where
 
 dividendos :: Int -> Double -> Double -> Double
-dividendos numAcoes precoAcao divYeld = numAcoes * precoAcao * divYeld
+dividendos numAcoes precoAcao divYeld = fromIntegral numAcoes * precoAcao * divYeld
